Update to top-of-tree glslang (2018-04-02). (#470)
This is an attempt to alleviate some driver crashes caused by invalid SPIR-V. Because Slang drives `glslang` with GLSL source code, any invalid output is likely due to `glslang` bugs.

I chose top-of-tree for `glslang` because it wasn't clear what their release process is. Hopefully we can go another year without having to update this dependency.

The build setup we use for `glslang` had to change to account for one more `#define` that the code expects to have passed in externally.