Pass AST interpolation modifiers through to codegen.
This is a short-term fix, because we (1) don't have an IR-level representation of interpolation qualifiers, and (2) can't introduce one until *after* the IR-level type system is introduced (to be able to handle `struct` fields).

The approach here is to find the AST-level declaration, either from layout information (in the case of an ordinary variable or function parameter), or from struct field information (because structs are being output from the AST form anyway).

I've included a single end-to-end rendering test to confirm that we handle the `nointerpolation` modifier the same as HLSL.

I also added the `noperspective` modifier, which seemed to be missing from our implementation.