Add an EditorConfig file (#474)
Developers may need to install an appropriate plugin for their text editor.
For example, Visual Studio 2015 users can install the plugin from `Tools -> Extensions and Updates...`.
Visual Studio 2017 supports EditorConfig directly.

The main goal of this PR is to allow VS users to more easily toggle between projects that default to spaces or tabs for indentation without having to change their VS options.

One side effect, however, is that EditorConfig might get overzealous and replace tabs with spaces in any file that is edited, so that PRs may contain spurious changes for a bit. The best fix for this would be to check in some project-wide fixup changes for formatting.