Allow implicit vector truncation (with a warning)
Fixes #502

The main change here is to add initializers ("constructors") to `vector<T,N>` types that take a `vector<T,M>` where `M > N`, and mark those as valid implicit conversions.

I also added logic to the semantic checker to emit a warning whenever code performs implicit conversion using an initializer that has a cost over a certain level. This ensures that the user will get warned if they do an unsafe implicit conversion.

As one more fix, I noticed that `uint` to `int` conversion was generating the warning, because we were treating it as a "general" conversion, but there are a lot of places that rely on that conversion so warning would create problems. I changed the cost of same-size signed-to-unsigned conversations to be lower, and actually lower than integer-to-float conversions.

When testing, I cound that one of our existing test cases was relying on implicit `float`-to-`int` conversion and now generates a warning. That seems reasonable, and I fixed the test case to do an explicit conversion instead.